Fix build and codestyle

+++ src/Shabalov/Task-3/Object3D.cpp
@@ -19,13 +19,14 @@ void Object3D::init(const std::vector<VertexData> &vertData,
 
   m_vertexBuffer.create();
   m_vertexBuffer.bind();
-  m_vertexBuffer.allocate(vertData.data(),
-                          vertData.size() * sizeof(VertexData));
+  m_vertexBuffer.allocate(vertData.data(), static_cast<int>(vertData.size()) *
+                                               sizeof(VertexData));
   m_vertexBuffer.release();
 
   m_indexBuffer.create();
   m_indexBuffer.bind();
-  m_indexBuffer.allocate(indexes.data(), indexes.size() * sizeof(GLuint));
+  m_indexBuffer.allocate(indexes.data(),
+                         static_cast<int>(indexes.size()) * sizeof(GLuint));
   m_indexBuffer.release();
 
   m_texture = std::make_unique<QOpenGLTexture>(texture.mirrored());
